BACKLOG-20105 Reorganize
