+++ src/javascript/ContentEditor/EditPanel/HeaderBadges/ReadOnlyBadge.scss
@@ -1,3 +1,3 @@
 .badge.badge {
-  margin-left: var(--spacing-small);
+    margin-left: var(--spacing-small);
 }
